requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
psycopg2-binary==2.9.1
pyshp==2.1.3
pyTelegramBotAPI==4.9.0
requests==2.26.0
requests-toolbelt==0.9.1
urllib3==1.26.7
python-dotenv
boto3
crc32c
fastapi
uvicorn
moto

